Added option to allow searching with offset and set limit of items returned. Also refactored to create searchService
