+++ lib/spotify-wrapper.js
@@ -1,17 +1,19 @@
 "use strict";
 
 const authService = require("./auth-service");
-const RequestBuilder = require("./requestBuilder");
-const MAX_ERROR = 3;
+const searchService = require("./search-service")
 
 // using an IIFE as you only need one instance of spotify api
 const SpotifyApi = (() => {
+
+  const MAX_ERROR = 3;
+  let err_cnt = 0;
+
   let clientID = "";
   let secretKey = "";
 
-  let token = "";
   let timeout_timestamp = "";
-  let err_cnt = 0;
+  let token = "";
 
   //takes in id and keys and call auth service to get token and timeout based on current time
   async function authenticate() {
@@ -37,25 +39,14 @@ const SpotifyApi = (() => {
   }
 
   //take query and searches against tracks
-  async function searchTrack(query) {
+  async function searchTrack(query, offset=0, limit=20) {
     try {
       //only auth if token has expired
-      if (Date.now() > timeout_timestamp) await authenticate();
+      if (!token || Date.now() > Date.parse(timeout_timestamp)) await authenticate();
 
-      const searchURL = `https://api.spotify.com/v1/search?q=${encodeURIComponent(
-        query
-      )}&type=track`;
-      const header = {
-        Authorization: "Bearer " + token,
-      };
-
-      //Build a request using builder and execute to send request
-      const request = new RequestBuilder();
-      const search = await request
-        .setUrl(searchURL)
-        .setMethod("get")
-        .setHeaders(header)
-        .execute();
+      //call our search service
+      const encodedQuery = `q=${encodeURIComponent(query)}&type=track&offset=${offset}&limit=${limit}`;
+      const search = await searchService(encodedQuery,token);
 
       err_cnt = 0;
       return search.data;
